services/storage/requirements.txt
aiohttp==3.3.2
motor==2.0.0
pytest==3.6.3
pytest-aiohttp==0.3.0
Pillow==5.2.0
